+++ lib/ORM/Relation.php
@@ -167,7 +167,7 @@ abstract class Relation extends BaseQuery
         }
 
         return $this->buildQuery()->toUpdate($this->connection)->update(array(
-            'deleted_at' => date($this->compiler->getDateFormat()),
+            'deleted_at' => date($this->model->getDateFormat()),
         ), true);
     }