+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 CHANGELOG
 -----------
+
+* Added support for soft deletes
+* `Opis\Database\Model::softDelete`, `Opis\Database\Model::supportsSoftDeletes`, `Opis\Database\ORM\Query::softDelete`,
+`Opis\Database\ORM\BaseQuery::withSoftDeleted`, `Opis\Database\ORM\BaseQuery::onlySoftDeleted` methods were added.
+
 ### v3.4.2, 2016.02.23
 
 * Fixed a bug in `Opis\Database\Model::__set` method