+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
 * Now you can count, update, delete, soft delete and restore related models
 * Added support for timestamps
 * Added `timestamps` and `softDelete` methods to `Opis\Database\Schema\CreateTable`
+* `Opis\Database\Model::getDateFormat` method is now public
 
 ### v3.4.2, 2016.02.23